fix(pkg): Replace 'localinstall' with 'install' for RPM package installation
The 'localinstall' command was deprecated and removed in dnf 5.x, as it was
only ever an alias for the 'install' command. This change uses 'install'
for all RPM-based systems, which works correctly with both older and newer
versions of dnf and yum.

Tested on Fedora 40 (dnf 4.x) and Fedora 41 (dnf 5.x).

Signed-off-by: Gavin Didrichsen <gavin.didrichsen@gmail.com>